rename and split connectors
