+++ src/api/collections-v4/collections.v4.controller.js
@@ -3,7 +3,7 @@ const {
   getAllRegistrationsByCouncils,
   getUnifiedRegistrations,
   updateRegistrationCollectedByCouncil
-} = require("../../connectors/registrationsDb/registrationsDb.connector");
+} = require("../../connectors/collectionsDb/collectionsDb.connector");
 
 const { validateOptions } = require("../collections/collections.service");
 const { transformRegForCollections } = require("../../services/collectionsTransform.service");
