+++ src/connectors/collectionsDb/collectionsDb.connector.js
@@ -2,7 +2,7 @@ const { establishConnectionToCosmos } = require("../cosmos.client");
 const { logEmitter } = require("../../services/logging.service");
 
 const getRegistrationsByCouncils = async (councils, collected, before, after) => {
-  logEmitter.emit("functionCall", "registrationsDb.connector", "getRegistrationsByCouncils");
+  logEmitter.emit("functionCall", "collectionsDb.connector", "getRegistrationsByCouncils");
   try {
     const registrationsCollection = await establishConnectionToCosmos(
       "registrations",
@@ -24,17 +24,17 @@ const getRegistrationsByCouncils = async (councils, collected, before, after) =>
       )
       .toArray();
 
-    logEmitter.emit("functionSuccess", "registrationsDb.connector", "getRegistrationsByCouncils");
+    logEmitter.emit("functionSuccess", "collectionsDb.connector", "getRegistrationsByCouncils");
 
     return registrations;
   } catch (err) {
-    logEmitter.emit("functionFail", "registrationsDb.connector", "getRegistrationsByCouncils", err);
+    logEmitter.emit("functionFail", "collectionsDb.connector", "getRegistrationsByCouncils", err);
     throw err;
   }
 };
 
 const getAllRegistrations = async (before, after) => {
-  logEmitter.emit("functionCall", "registrationsDb.connector", "getAllRegistrations");
+  logEmitter.emit("functionCall", "collectionsDb.connector", "getAllRegistrations");
   try {
     let registrationsCollection = await establishConnectionToCosmos(
       "registrations",
@@ -53,16 +53,16 @@ const getAllRegistrations = async (before, after) => {
         { projection: { _id: 0, "fsa-rn": 1 } }
       )
       .toArray();
-    logEmitter.emit("functionSuccess", "registrationsDb.connector", "getAllRegistrations");
+    logEmitter.emit("functionSuccess", "collectionsDb.connector", "getAllRegistrations");
     return registrations;
   } catch (err) {
-    logEmitter.emit("functionFail", "registrationsDb.connector", "getAllRegistrations", err);
+    logEmitter.emit("functionFail", "collectionsDb.connector", "getAllRegistrations", err);
     throw err;
   }
 };
 
 const getFullRegistration = async (fsa_rn, fields = []) => {
-  logEmitter.emit("functionCall", "registrationsDb.connector", "getFullRegistration");
+  logEmitter.emit("functionCall", "collectionsDb.connector", "getFullRegistration");
   try {
     const projection = Object.assign(
       {
@@ -94,17 +94,17 @@ const getFullRegistration = async (fsa_rn, fields = []) => {
       { projection: projection }
     );
 
-    logEmitter.emit("functionSuccess", "registrationsDb.connector", "getFullRegistration");
+    logEmitter.emit("functionSuccess", "collectionsDb.connector", "getFullRegistration");
 
     return registration;
   } catch (err) {
-    logEmitter.emit("functionFail", "registrationsDb.connector", "getFullRegistration", err);
+    logEmitter.emit("functionFail", "collectionsDb.connector", "getFullRegistration", err);
     throw err;
   }
 };
 
 const getSingleRegistration = async (fsa_rn, council) => {
-  logEmitter.emit("functionCall", "registrationsDb.connector", "getSingleRegistration");
+  logEmitter.emit("functionCall", "collectionsDb.connector", "getSingleRegistration");
 
   const projection = Object.assign({
     _id: 0,
@@ -133,16 +133,16 @@ const getSingleRegistration = async (fsa_rn, council) => {
   if (registration === null) {
     const error = new Error("getRegistrationNotFoundError");
     error.name = "getRegistrationNotFoundError";
-    logEmitter.emit("functionFail", "registrationsDb.connector", "getSingleRegistration", error);
+    logEmitter.emit("functionFail", "collectionsDb.connector", "getSingleRegistration", error);
     throw error;
   }
-  logEmitter.emit("functionSuccess", "registrationsDb.connector", "getSingleRegistration");
+  logEmitter.emit("functionSuccess", "collectionsDb.connector", "getSingleRegistration");
 
   return registration;
 };
 
 const getUnifiedRegistrations = async (before, after) => {
-  logEmitter.emit("functionCall", "registrationsDb.connector", "getUnifiedRegistrations");
+  logEmitter.emit("functionCall", "collectionsDb.connector", "getUnifiedRegistrations");
 
   // convert ISOStrings to Date type
   const beforeDate = new Date(before);
@@ -155,12 +155,12 @@ const getUnifiedRegistrations = async (before, after) => {
     })
   );
 
-  logEmitter.emit("functionSuccess", "registrationsDb.connector", "getUnifiedRegistrations");
+  logEmitter.emit("functionSuccess", "collectionsDb.connector", "getUnifiedRegistrations");
   return fullRegistrations;
 };
 
 const getAllRegistrationsByCouncils = async (councils, newRegistrations, fields, before, after) => {
-  logEmitter.emit("functionCall", "registrationsDb.connector", "getAllRegistrationsByCouncils");
+  logEmitter.emit("functionCall", "collectionsDb.connector", "getAllRegistrationsByCouncils");
 
   // get NEW [false, null] or EVERYTHING [true, false, null]
   const queryArray = newRegistrations === "true" ? [false] : [true, false];
@@ -171,14 +171,14 @@ const getAllRegistrationsByCouncils = async (councils, newRegistrations, fields,
       return getFullRegistration(registration["fsa-rn"], fields);
     })
   );
-  logEmitter.emit("functionSuccess", "registrationsDb.connector", "getAllRegistrationsByCouncils");
+  logEmitter.emit("functionSuccess", "collectionsDb.connector", "getAllRegistrationsByCouncils");
   return fullRegistrations;
 };
 
 const updateRegistrationCollectedByCouncil = async (fsa_rn, collected, council) => {
   logEmitter.emit(
     "functionCall",
-    "registrationsDb.connector",
+    "collectionsDb.connector",
     "updateRegistrationCollectedByCouncil"
   );
 
@@ -202,7 +202,7 @@ const updateRegistrationCollectedByCouncil = async (fsa_rn, collected, council) 
     error.name = "updateRegistrationNotFoundError";
     logEmitter.emit(
       "functionFail",
-      "registrationsDb.connector",
+      "collectionsDb.connector",
       "updateRegistrationCollectedByCouncil",
       error
     );
@@ -210,7 +210,7 @@ const updateRegistrationCollectedByCouncil = async (fsa_rn, collected, council) 
   }
   logEmitter.emit(
     "functionSuccess",
-    "registrationsDb.connector",
+    "collectionsDb.connector",
     "updateRegistrationCollectedByCouncil"
   );
   return { fsa_rn, collected };
