+++ src/connectors/submissionsDb/submissionsDb.connector.test.js
@@ -1,9 +1,9 @@
 jest.mock("mongodb");
 
 const mongodb = require("mongodb");
-const { cacheRegistration, updateNotificationOnSent } = require("./cacheDb.connector");
+const { cacheRegistration, updateNotificationOnSent } = require("./submissionsDb.connector");
 
-describe("Connector: cacheDb", () => {
+describe("Connector: submissionsDb", () => {
   let response;
 
   beforeEach(() => {
@@ -82,74 +82,6 @@ describe("Connector: cacheDb", () => {
 
       it("should have used the first mongo connnection both times", () => {
         expect(response.insertedId).toBe("1000");
-      });
-    });
-  });
-
-  describe("Function: updateCompletedInCache1", () => {
-    describe("Function: updateNotificationOnSent", () => {
-      describe("When success", () => {
-        let result;
-        let testStatus = {
-          notifications: [
-            {
-              type: "LC",
-              address: "fsatestemail.valid@gmail.com",
-              time: undefined,
-              sent: undefined
-            }
-          ]
-        };
-
-        let testEmailsToSend = [
-          {
-            type: "LC",
-            address: "fsatestemail.valid@gmail.com",
-            templateId: "testtempate234315431asdfasf"
-          }
-        ];
-
-        it("check it creates notification rows correctly", () => {
-          result = updateNotificationOnSent(
-            testStatus,
-            "FAKE-FSAID-12345",
-            testEmailsToSend,
-            0,
-            true,
-            "fakedate"
-          );
-
-          expect(result).toStrictEqual({
-            notifications: [
-              {
-                address: "fsatestemail.valid@gmail.com",
-                type: "LC",
-                sent: true,
-                time: "fakedate"
-              }
-            ]
-          });
-
-          result = updateNotificationOnSent(
-            testStatus,
-            "FAKE-FSAID-12345",
-            testEmailsToSend,
-            0,
-            false,
-            "fakedate"
-          );
-
-          expect(result).toStrictEqual({
-            notifications: [
-              {
-                address: "fsatestemail.valid@gmail.com",
-                type: "LC",
-                sent: false,
-                time: "fakedate"
-              }
-            ]
-          });
-        });
       });
     });
   });
